fixed bug: generateEvent() method
The EventSpecs object used an incorrect/incompatible property "canBubble" from legacy versions.  The generateEvent() method read the property from the specs, and passed it as an argument to the Event.initEvent() method.  Newer browsers can accept this EventSpecs object directly using Event constructors, and the property must be "bubbles" instead of "canBubble".  This version of UniDOM requires that the "specs" argument for generateEvent() now recognizes/uses the property "bubbles", not "canBubble".  The previous version could cause a bug in the coded algorithm when used in newer browsers that have Event constructors, when using the "canBubble" property.